Title: 세 수의 중간값, Time: 72ms, Memory: 18MB, Status: Runtime Error - CodeTree

+++ 250108/세 수의 중간값/median-of-three-numbers.py
@@ -1,3 +1,3 @@
-a, b, c = float(input().split())
+a, b, c = int(input().split())
 
 print(1) if (b>a and b<c) else print(0)
